+++ bot/botconfig/config.json
@@ -1,5 +1,4 @@
 {
-  "loadSlashsGlobal": true, "_COMMENT_": "Set it to true, when your Slash Commands are ready, false loads instant true loads slow but stable! (GLOBAL/GUILD COMMANDS, DISCORD DOCS!)",
   "slashCommandsDirs": [
     {
       "Folder": "suggestions",
