+++ bot/events/client/disconnect.js
@@ -1,5 +1,5 @@
 //here the event starts
-const Logger = require("../../handlers/logger");
+const Logger = require("../../../logger");
 module.exports = client => {
     Logger.info(`Disconnected at ${new Date()}.`);
 }
