+++ bot/events/client/error.js
@@ -1,5 +1,5 @@
 //here the event starts
-const Logger = require("../../handlers/logger");
+const Logger = require("../../../logger");
 module.exports = (client, error) => {
     Logger.error(`${error}`);
 }
