+++ bot/events/client/rateLimit.js
@@ -1,5 +1,5 @@
 //here the event starts
-const Logger = require("../../handlers/logger");
+const Logger = require("../../../logger");
 module.exports = (client, rateLimitData) => {
     Logger.super_error(`${rateLimitData.method} ${rateLimitData.path}`, `${rateLimitData}`);
 }
