+++ bot/events/client/ready.js
@@ -1,7 +1,7 @@
 //here the event starts
 const config = require("../../botconfig/config.json")
 const {change_status} = require("../../handlers/functions");
-const Logger = require("../../handlers/logger");
+const Logger = require("../../../logger");
 module.exports = client => {
     //SETTING ALL GUILD DATA FOR THE DJ ONLY COMMANDS for the DEFAULT
     //client.guilds.cache.forEach(guild=>client.settings.set(guild.id, ["autoplay", "clearqueue", "forward", "loop", "jump", "loopqueue", "loopsong", "move", "pause", "resume", "removetrack", "removedupe", "restart", "rewind", "seek", "shuffle", "skip", "stop", "volume"], "djonlycmds"))
