+++ bot/events/client/shardDisconnect.js
@@ -1,5 +1,5 @@
 //here the event starts
-const Logger = require("../../handlers/logger");
+const Logger = require("../../../logger");
 module.exports = (client, event, id) => {
     Logger.error(`[${String(new Date).split(" ", 5).join(" ")}] Shard #${id} Disconnected`);
 }
