+++ bot/events/client/shardError.js
@@ -1,5 +1,5 @@
 //here the event starts
-const Logger = require("../../handlers/logger");
+const Logger = require("../../../logger");
 module.exports = (client, error, id) => {
     Logger.error(`[${String(new Date).split(" ", 5).join(" ")}] Shard #${id} Errored:`, error);
 }
