+++ bot/events/client/shardReady.js
@@ -1,5 +1,5 @@
 //here the event starts
-const Logger = require("../../handlers/logger");
+const Logger = require("../../../logger");
 module.exports = (client, id) => {
     Logger.info(`[${String(new Date).split(" ", 5).join(" ")}] Shard #${id} Ready`);
 }
