+++ bot/events/guild/messageReactionRemove.js
@@ -12,7 +12,7 @@ module.exports = async (client, reaction, user) => {
     if (reaction.message.partial) await reaction.message.fetch();
     if (reaction.partial) await reaction.fetch();
     if (user.bot) return;
-    if (!reaction.message.guild){
+    if (!reaction.message.guild) {
         //REST OF THE CODE HERE
     }
 }
