+++ bot/events/guild/threadCreate.js
@@ -1,4 +1,4 @@
-const Logger = require("../../handlers/logger");
+const Logger = require("../../../logger");
 module.exports = async (client, thread) => {
     if (thread.joinable) {
         try {
