+++ bot/slashCommands/suggest.js
@@ -1,10 +1,10 @@
-const Logger = require("../../../logger");
+const Logger = require("../../logger");
 const modals = require("discord-modals");
-const db = require("../../../database");
+const db = require("../../database");
 
 module.exports = {
-    name: "add", //the command name for the Slash Command
-    description: "Add a suggestion.  Same as /suggest", //the command description for Slash Command Overview
+    name: "suggest", //the command name for the Slash Command
+    description: "Add a suggestion. Same as /suggestions add", //the command description for Slash Command Overview
     cooldown: 5, //the cooldown for the command in seconds (max. 60)
     memberpermissions: [], //Only allow members with specific Permissions to execute a Commmand [OPTIONAL]
     requiredroles: [], //Only allow specific Users with a Role to execute a Command [OPTIONAL]
@@ -14,7 +14,7 @@ module.exports = {
         try {
 
             const language = await db.getServerLanguage(interaction.guild.id)
-            const lang = require(`../../botconfig/languages/${language}.json`);
+            const lang = require(`../botconfig/languages/${language}.json`);
 
             const modal = new modals.Modal()
                 .setCustomId('send')
