+++ bot/startbot.js
@@ -1,4 +1,5 @@
 const Discord = require("discord.js");
+const logger
 const client = new Discord.Client({
 
     //fetchAllMembers: false,
@@ -47,7 +48,7 @@ client.aliases = new Discord.Collection();
 
 //Require the Handlers                  Add the antiCrash file too, if its enabled
 
-["events", "commands", "slashCommands", process.env.BOT_ANTI_CRASH ? "antiCrash" : null]
+["events", /*"commands",*/ "slashCommands", process.env.BOT_ANTI_CRASH ? "antiCrash" : null]
     .filter(Boolean)
     .forEach(h => {
         require(`./handlers/${h}`)(client);
